+++ hooks/ocean-watch/index.js
@@ -1,22 +1,15 @@
-import {
-  useQuery,
-} from 'react-query';
+import { useQuery } from 'react-query';
 
 // services
-import {
-  fetchOceanWatchAreas,
-} from 'services/ocean-watch';
+import { fetchOceanWatchAreas } from 'services/ocean-watch';
 
-export const useOceanWatchAreas = (queryConfig = {}) => useQuery(
-  ['ocean-watch-areas'],
-  () => fetchOceanWatchAreas(),
-  {
+export const useOceanWatchAreas = (queryConfig = {}) =>
+  useQuery(['ocean-watch-areas'], () => fetchOceanWatchAreas(), {
     refetchOnWindowFocus: false,
     placeholderData: [],
     initialStale: true,
     ...queryConfig,
-  },
-);
+  });
 
 export default {
   useOceanWatchAreas,