+++ pages/dashboards/ocean-watch/partners/index.jsx
@@ -1,11 +1,6 @@
-import {
-  useMemo,
-  useCallback,
-} from 'react';
+import { useMemo, useCallback } from 'react';
 import { useRouter } from 'next/router';
-import {
-  useQueryClient,
-} from 'react-query';
+import { useQueryClient } from 'react-query';
 
 // components
 import LayoutOceanWatch from 'layout/layout/ocean-watch';
@@ -15,12 +10,8 @@ import PartnerBlock from 'components/partner-block';
 import BannerCountries from 'components/banners/countries';
 
 // hooks
-import {
-  usePublishedPartners,
-} from 'hooks/partners';
-import {
-  useOceanWatchAreas,
-} from 'hooks/ocean-watch';
+import { usePublishedPartners } from 'hooks/partners';
+import { useOceanWatchAreas } from 'hooks/ocean-watch';
 
 const PARTNERS_PAGE_DESCRIPTION = `
 We couldn’t do this on our own.
@@ -34,114 +25,111 @@ export default function OceanWatchPartnersPage() {
   const router = useRouter();
   const queryClient = useQueryClient();
   const {
-    data: {
-      collaboratingPartners,
-      dataProviders,
-      funders,
+    data: { collaboratingPartners, dataProviders, funders },
+  } = usePublishedPartners(
+    {},
+    {
+      select: (_partners) => ({
+        collaboratingPartners: _partners.filter(
+          (_partner) => _partner['partner-type'] === 'ow_collaborating-partner',
+        ),
+        dataProviders: _partners.filter(
+          (_partner) => _partner['partner-type'] === 'ow_data-provider',
+        ),
+        funders: _partners.filter((_partner) => _partner['partner-type'] === 'ow_funder'),
+      }),
+      placeholderData: [],
+      refetchOnWindowFocus: false,
     },
-  } = usePublishedPartners({}, {
-    select: (_partners) => ({
-      collaboratingPartners: _partners.filter((_partner) => _partner['partner-type'] === 'ow_collaborating-partner'),
-      dataProviders: _partners.filter((_partner) => _partner['partner-type'] === 'ow_data-provider'),
-      funders: _partners.filter((_partner) => _partner['partner-type'] === 'ow_funder'),
-    }),
-    placeholderData: [],
-    refetchOnWindowFocus: false,
-  });
+  );
 
-  const {
-    data: countries,
-  } = useOceanWatchAreas({
+  const { data: countries } = useOceanWatchAreas({
     placeholderData: queryClient.getQueryData('ocean-watch-areas') || [],
-    select: (_countries) => _countries.map(({ name, iso }) => ({
-      label: name,
-      value: iso,
-    })),
+    select: (_countries) =>
+      _countries
+        .filter(({ iso }) => iso !== 'GLB')
+        .map(({ name, iso }) => ({
+          label: name,
+          value: iso,
+        })),
     refetchOnWindowFocus: false,
   });
 
-  const handleCountry = useCallback((iso) => {
-    router.push({
-      pathname: '/dashboards/ocean-watch/country/[iso]',
-      query: {
-        iso,
+  const handleCountry = useCallback(
+    (iso) => {
+      router.push({
+        pathname: '/dashboards/ocean-watch/country/[iso]',
+        query: {
+          iso,
+        },
+      });
+    },
+    [router],
+  );
+
+  const partners = useMemo(
+    () => [
+      {
+        title: 'Collaborating partners',
+        children: collaboratingPartners,
       },
-    });
-  }, [router]);
-
-  const partners = useMemo(() => [
-    {
-      title: 'Collaborating partners',
-      children: collaboratingPartners,
-    },
-    {
-      title: 'Data Providers',
-      children: dataProviders,
-    },
-    {
-      title: 'Funders',
-      children: funders,
-    },
-
-  ], [collaboratingPartners, dataProviders, funders]);
+      {
+        title: 'Data Providers',
+        children: dataProviders,
+      },
+      {
+        title: 'Funders',
+        children: funders,
+      },
+    ],
+    [collaboratingPartners, dataProviders, funders],
+  );
 
   return (
-    <LayoutOceanWatch
-      title="Ocean Watch – Partners"
-      description={PARTNERS_PAGE_DESCRIPTION}
-    >
+    <LayoutOceanWatch title="Ocean Watch – Partners" description={PARTNERS_PAGE_DESCRIPTION}>
       <Header className="-transparent" />
       <OceanWatchHero className="-ocean-watch" />
       <section className="l-section -secondary -medium">
         <div className="l-container">
           <div className="row">
             <div className="column small-12 medium-8">
-              <h2>
-                Managing an integrated ocean requires collaboration
-              </h2>
-              <p>
-                {PARTNERS_PAGE_DESCRIPTION}
-              </p>
+              <h2>Managing an integrated ocean requires collaboration</h2>
+              <p>{PARTNERS_PAGE_DESCRIPTION}</p>
             </div>
           </div>
         </div>
       </section>
-      {(partners.map(({ title, children }) => (children.length > 0) && (
-        <section
-          key={title}
-          className="l-section"
-        >
-          <div className="l-container">
-            <div className="row">
-              <div className="column small-12">
-                <h2 className="-text-center">
-                  {title}
-                </h2>
-              </div>
-            </div>
-            <div className="row">
-              {children.map((child) => (
-                <div
-                  key={child.id}
-                  className="column small-12 medium-6"
-                >
-                  <PartnerBlock item={child} />
+      {partners.map(
+        ({ title, children }) =>
+          children.length > 0 && (
+            <section key={title} className="l-section">
+              <div className="l-container">
+                <div className="row">
+                  <div className="column small-12">
+                    <h2 className="-text-center">{title}</h2>
+                  </div>
                 </div>
-              ))}
-            </div>
-          </div>
-        </section>
-      )))}
+                <div className="row">
+                  {children.map((child) => (
+                    <div key={child.id} className="column small-12 medium-6">
+                      <PartnerBlock item={child} />
+                    </div>
+                  ))}
+                </div>
+              </div>
+            </section>
+          ),
+      )}
       <section className="l-section -small">
         <div className="l-container">
           <BannerCountries
-            title={(
+            title={
               <>
                 Explore the Ocean Watch
                 <br />
                 local data
               </>
-              )}
+            }
             onChangeCountry={handleCountry}
             countryList={countries}
           />