+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - custom "no data" for set of indicators. [OW-179](https://vizzuality.atlassian.net/browse/OW-179)
 
 ### Changed
+- ocean watch: removes `the World` in country index page.
+- ocean watch: removes `the World` from coastline options list.
 - moves native `scripts` to NextJS scripts.
 - updates `eslint@8.9.0` and other linters and plugins.
 - `next@12.1.0`
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `next@12.0.10`
 
 ### Fixed
+- ocean watch: highlight of "Coastal Profile" tab.
 - highlighting of navigation bar based on current page. [RW-92](https://vizzuality.atlassian.net/browse/RW-92)
 - storybook: star icon appearing in widget header. [RW-78](https://vizzuality.atlassian.net/browse/RW-78)
 - wrong attributes sent to the user data creation endpoint.